Use `double` for all percentages
Output now looks like this:

$ ./earlyoom
earlyoom v1.4-18-gfb4205d-dirty
mem total: 23909 MiB, swap total:    0 MiB
sending SIGTERM when mem <= 10.00% and swap <= 10.00%,
        SIGKILL when mem <=  5.00% and swap <=  5.00%
Could not lock memory - continuing anyway: Cannot allocate memory
mem avail: 20100 of 23909 MiB (84.07%), swap free:    0 of    0 MiB ( 0.00%)
mem avail: 20110 of 23909 MiB (84.11%), swap free:    0 of    0 MiB ( 0.00%)

+++ kill.h
@@ -8,10 +8,10 @@
 typedef struct {
     /* if the available memory AND swap goes below these percentages,
      * we start killing processes */
-    int mem_term_percent;
-    int mem_kill_percent;
-    int swap_term_percent;
-    int swap_kill_percent;
+    double mem_term_percent;
+    double mem_kill_percent;
+    double swap_term_percent;
+    double swap_kill_percent;
     /* ignore /proc/PID/oom_score_adj? */
     bool ignore_oom_score_adj;
     /* notifcation command to launch when killing something. NULL = no-op. */
